+++ src/test/java/com/faforever/client/game/GameRunnerTest.java
@@ -12,7 +12,9 @@ import com.faforever.client.domain.server.GameInfo;
 import com.faforever.client.domain.server.PlayerInfo;
 import com.faforever.client.fa.ForgedAllianceLaunchService;
 import com.faforever.client.fa.GameParameters;
-import com.faforever.client.fa.relay.ice.IceAdapter;
+import com.faforever.client.fa.relay.gpg.GPGNetServer;
+import com.faforever.client.fa.relay.gpg.LobbyInitMode;
+import com.faforever.client.fa.relay.ice.IceAdapterService;
 import com.faforever.client.featuredmod.FeaturedModService;
 import com.faforever.client.fx.FxApplicationThreadExecutor;
 import com.faforever.client.fx.ObservableConstant;
@@ -103,7 +105,7 @@ public class GameRunnerTest extends ServiceTest {
   @Mock
   private ReplayServer replayServer;
   @Mock
-  private IceAdapter iceAdapter;
+  private GPGNetServer gpgNetServer;
   @Mock
   private ModService modService;
   @Mock
@@ -130,6 +132,8 @@ public class GameRunnerTest extends ServiceTest {
   private FxApplicationThreadExecutor fxApplicationThreadExecutor;
   @Mock
   private GamePathHandler gamePathHandler;
+  @Mock
+  private IceAdapterService iceAdapterService;
   @Spy
   private GameMapper gameMapper = Mappers.getMapper(GameMapper.class);
   @Spy
@@ -158,9 +162,10 @@ public class GameRunnerTest extends ServiceTest {
     lenient().when(preferencesService.hasValidGamePath()).thenReturn(true);
     lenient().when(fafServerAccessor.connectionStateProperty()).thenReturn(new SimpleObjectProperty<>());
     lenient().when(replayServer.start(anyInt())).thenReturn(completedFuture(LOCAL_REPLAY_PORT));
-    lenient().when(iceAdapter.start(anyInt())).thenReturn(completedFuture(GPG_PORT));
+    lenient().when(gpgNetServer.start(anyInt(), any(LobbyInitMode.class))).thenReturn(completedFuture(GPG_PORT));
     lenient().when(playerService.getCurrentPlayer()).thenReturn(junitPlayer);
     lenient().when(process.pid()).thenReturn(10L);
+    lenient().when(iceAdapterService.getNewest()).thenReturn(Mono.empty());
 
     lenient().doAnswer(invocation -> {
       try {
@@ -183,7 +188,7 @@ public class GameRunnerTest extends ServiceTest {
     }
     lenient().when(forgedAllianceLaunchService.launchOnlineGame(any(), anyInt(), anyInt())).thenReturn(process);
     lenient().when(replayServer.start(anyInt())).thenReturn(completedFuture(LOCAL_REPLAY_PORT));
-    lenient().when(iceAdapter.start(anyInt())).thenReturn(completedFuture(GPG_PORT));
+    lenient().when(gpgNetServer.start(anyInt(), any(LobbyInitMode.class))).thenReturn(completedFuture(GPG_PORT));
     lenient().when(process.onExit()).thenReturn(new CompletableFuture<>());
     lenient().when(process.exitValue()).thenReturn(0);
     lenient().when(process.isAlive()).thenReturn(true);
@@ -212,7 +217,7 @@ public class GameRunnerTest extends ServiceTest {
     verify(leaderboardService, never()).getActiveLeagueEntryForPlayer(any(), any());
     verify(mapService, never()).downloadIfNecessary(any());
     verify(replayServer).start(uid);
-    verify(iceAdapter).start(uid);
+    verify(gpgNetServer).start(uid, LobbyInitMode.NORMAL);
     assertTrue(instance.isRunning());
     assertEquals(uid, instance.getRunningGame().getId());
     assertEquals(10L, instance.getRunningProcessId());
@@ -222,7 +227,7 @@ public class GameRunnerTest extends ServiceTest {
 
     exitFuture.complete(process);
 
-    verify(iceAdapter).stop();
+    verify(gpgNetServer).stop();
     verify(replayServer).stop();
     verify(fafServerAccessor).notifyGameEnded();
     verify(fafServerAccessor).setPingIntervalSeconds(25);
@@ -353,7 +358,7 @@ public class GameRunnerTest extends ServiceTest {
 
     testNoticePublisher.next(new NoticeInfo("kill", "test"));
 
-    verify(iceAdapter).stop();
+    verify(gpgNetServer).stop();
     verify(process).destroy();
     verify(notificationService).addNotification(any(ImmediateNotification.class));
   }
